clicking on the storyboard from the dashboard leads to the storyboard page with the story loaded
