+++ src/lib/stores/storyboard.ts
@@ -1,0 +1,4 @@
+import { writable } from 'svelte/store';
+import type { Storyboard } from '$lib/models/storyboard.model';
+
+export const storyboardStore = writable<Storyboard | null>(null);
